add blue buck in dashboard and referral event in ag
