+++ projects/navi-aggregator/sentio.yaml
@@ -2,3 +2,5 @@ project: navi/dex-aggregator
 contracts:
   - chain: sui_mainnet
     address: "0x88dfe5e893bc9fa984d121e4d0d5b2e873dc70ae430cf5b3228ae6cb199cb32b"
+  - chain: sui_mainnet
+    address: "0x5b7d732adeb3140a2dbf2becd1e9dbe56cee0e3687379bcfe7df4357ea664313"
